Atualiza font

+++ front/src/pages/grupo/ListagemGrupos.tsx
@@ -1,3 +1,4 @@
+
 import { useEffect, useRef, useState } from "react";
 import axios from "axios";
 import autoAnimate from "@formkit/auto-animate";
@@ -5,31 +6,42 @@ import { Container, Table } from "react-bootstrap";
 import { FaSortUp, FaSortDown, FaChevronLeft, FaChevronRight } from "react-icons/fa";
 import ReactPaginate from "react-paginate";
 import { Link, useNavigate } from "react-router-dom";
-import Header from "../components/Header";
-import { avisoErroAoDeletar } from "../controllers";
-import { avisoDeletar } from "../controllers/avisoConcluido";
-import { avisoErroDeletar } from "../controllers/avisoErro";
-import { URIcommit, URIattach, URI, URIgroupToUser, URIgroup } from "../enumerations/uri";
-import { removeFile } from "../services/supabase";
-import { Calls } from "../types/call";
-import editar from "../images/editar.png";
-import excluir from "../images/excluir.png";
-import grupoImag from "../images/grupo.png";
-import "../App.css";
-import { GroupsToUser } from "../types/groupToUser";
-import { Groups } from "../types/group";
-import { group } from "console";
+import Header from "../../components/Header";
+import { avisoErroAoDeletar } from "../../controllers";
+import { avisoDeletar } from "../../controllers/avisoConcluido";
+import { avisoErroDeletar, avisoChamado } from "../../controllers/avisoErro";
+import { URIcommit, URIattach, URI, URIgroupToUser, URIgroup } from "../../enumerations/uri";
+import editar from "../../images/editar.png";
+import excluir from "../../images/excluir.png";
+import grupoImag from "../../images/grupo.png";
+import "../../App.css";
+import { GroupsToUser } from "../../types/groupToUser";
+import { Groups } from "../../types/group";
+
 
 function ListagemGrupos() {
 
   const url_atual = window.location.href;
   const id = window.location.href.split("/")[4]
 
+  const [dataGroup, setGroup] = useState<Groups[]>([]);
   const [data, setData] = useState<GroupsToUser[]>([]);
 
   //axios get
   useEffect(() => {
-    async function fetchCalls() {
+    async function fetchGroup() {
+      axios
+        .get(URIgroup.PEGAR_GROUP)
+        .then((response) => {
+          setGroup(response.data);
+        })
+        .catch((error) => {
+          console.log(error);
+        });
+    }
+
+    fetchGroup();
+    async function fetchGroupToUser() {
       axios
         .get(URIgroupToUser.PEGAR_GROUP_TO_USER)
         .then((response) => {
@@ -39,31 +51,42 @@ function ListagemGrupos() {
           console.log(error);
         });
     }
-    fetchCalls();
+    fetchGroupToUser();
 
   }, []);
 
-
+  console.log(data);
+  console.log(dataGroup);
+  
 
   //delete  
   async function handleDeleteGroupUser(id: number) {
     try {
       const groupToUserEntries = data.filter(dataGroup => dataGroup.group.id === id);
-
+      console.log(groupToUserEntries);
+      
       const shouldDelete = await avisoDeletar();
       if (shouldDelete.isConfirmed) {
-        await Promise.all(groupToUserEntries.map(async (groupToUserEntry) => {
-          const { id: groupToUserId, group: { id: groupId }, user: { id: userId } } = groupToUserEntry;
-          await axios.delete(`${URIgroupToUser.DELETE_GROUP_TO_USER}${groupToUserId}`);
-          console.log(`Entrada GrupoToUser ${groupToUserId} excluída`);
-        }));
+        if(dataGroup.find((item) => item.id == id && item.groupType == "Funcionario") !=undefined){
+          await Promise.all(groupToUserEntries.map(async (groupToUserEntry) => {
+            const { id: groupToUserId} = groupToUserEntry;
+            await axios.delete(`${URIgroupToUser.DELETE_GROUP_TO_USER}${groupToUserId}`);
+          }));
+        }
+        
+          await axios.delete(`${URIgroup.DELETE_GROUP}${id}`).then(() => {  
+            setTimeout(() => {
+            window.location.reload();
+          }, 200);})
+          .catch((err) => {
+           avisoChamado();
+          })
 
-        await axios.delete(`${URIgroup.DELETE_GROUP}${id}`);
-        console.log(`Grupo ${id} excluído`);
       }
+
     } catch (error) {
       console.error(error);
-      avisoErroDeletar();
+      avisoErroDeletar()
     }
   }
 
@@ -92,10 +115,8 @@ function ListagemGrupos() {
 
 
   let grupo: any = null;
-
   data.forEach((user) => {
     if (grupo === null) {
-      // Se este for o primeiro usuário, armazene as informações do grupo
       grupo = {
         id: user.group.id,
         usuarios: [],
@@ -104,7 +125,6 @@ function ListagemGrupos() {
       grupo.usuarios.push(user.user.userName);
       grupo.usuarioIDs.push(user.user.id);
     } else if (user.group.id === grupo.id) {
-      // Se o ID do grupo deste usuário for igual ao ID do grupo armazenado, adicione o nome do usuário ao array de usuários
       grupo.usuarios.push(user.user.userName);
       grupo.usuarioIDs.push(user.user.id);
       console.log(grupo.usuarioIDs);
@@ -124,7 +144,6 @@ function ListagemGrupos() {
     return result;
   }, {});
   const groupList = Object.values(groupedData);
-
   console.log(groupedData);
   
 
@@ -174,22 +193,23 @@ function ListagemGrupos() {
                 <thead>
                   <tr>
                     {/*cabeçalho tabela*/}
-                    <th onClick={() => sorting("id")} className="text-center">Nome da Equipe {order === "ASC" ? <FaSortUp /> : <FaSortDown />} </th>
+                    <th onClick={() => sorting("id")} className="text-center">Nome do Grupo {order === "ASC" ? <FaSortUp /> : <FaSortDown />} </th>
+                    <th className="text-center">Tipo do Grupo</th>
                     <th className="text-center">Ações</th>
                     {/*fim cabeçalho tabela*/}
                   </tr>
                 </thead>
 
                 <tbody>
-                  {groupList.map((grupo: any) => (
+                  {dataGroup.map((grupo: any) => (
                     <tr key={grupo.id}>
                       {/*corpo tabela*/}
+                      <td className="text-center">{grupo.groupName}</td>
                       <td className="text-center">{grupo.groupType}</td>
                       <td className="text-center">
-                        <Link to={"/editarGrupo/" + grupo.id} style={{ padding: "3px" }}>
+                        <Link to={"/editarGrupo/" + grupo.id + "/" + grupo.groupType} style={{ padding: "3px" }}>
                           <img src={editar} style={{ width: '25px' }} alt='Editar' />
                         </Link>
-                        {grupo.usuarios.length <= 5 ? (
                           <img
                             className="actions"
                             style={{ width: "35px", padding: "3px" }}
@@ -197,9 +217,6 @@ function ListagemGrupos() {
                             alt="Excluir"
                             onClick={() => handleDeleteGroupUser(grupo.id)}
                           />
-                        ) : (
-                          <span>{`${grupo.usuarios.length} usuários`}</span>
-                        )}
                       </td>
                     </tr>
                   ))}
